requirements.txt
absl-py==2.1.0
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli @ file:///croot/brotli-split_1714483155106/work
bs4==0.0.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
comet-ml==3.44.1
configobj==5.0.8
contourpy==1.2.1
cycler==0.12.1
docstring_parser==0.16
dulwich==0.22.1
einops==0.8.0
everett==3.1.0
filelock @ file:///croot/filelock_1700591183607/work
flash-attn==2.5.9.post1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.23.4
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lightning-utilities==0.11.3.post0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.1
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.5
mypy-extensions==1.0.0
networkx @ file:///croot/networkx_1720002482208/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.7.0
packaging==24.1
pandas==2.2.2
pillow @ file:///croot/pillow_1714398848491/work
protobuf==4.25.3
psutil==6.0.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-box==6.1.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests @ file:///croot/requests_1716902831423/work
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.0
safetensors==0.4.3
scipy==1.14.0
semantic-version==2.10.0
sentry-sdk==2.9.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.0.0
torchmetrics==1.4.0.post0
torchvision==0.15.0
tqdm==4.66.4
transformers==4.42.3
triton==2.3.1
typed-argument-parser==1.10.1
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1718912636303/work
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
wurlitzer==3.1.1
xformers==0.0.27
yarl==1.9.4
